Revert "refactor: manualPassportInput into InputPassportSection (#361)"
This reverts commit df112410f393b3cf3c0d65cbfe2a7a5399c653ae.
